Increase the lines searched for PNG image
Only effects plaintext gcode. The new QOI images make it necessary to
look further into the gcode to find the PNG image for display in
Prusalink.

In the interest of not wasting time checking when the correct size
definitely doesn't exist, we check to see if the current thumbnail is
larger than asked for. If it is, and we do not want larger thumbnails,
we end the search and report no thumbnail found. This works because the
thumbnails are ordered from smallest to largest.